factor DB page filters into a separate component
